float to int

+++ tests/unit/test_pg_conn_grace_period.py
@@ -1,11 +1,10 @@
 # encoding: utf-8
 from unittest.mock import patch
 
-
 from src.pg_conn_grace_period import PgConnGracePeriod
 
 
-def make(grace: float = 30.0) -> PgConnGracePeriod:
+def make(grace: int = 30) -> PgConnGracePeriod:
     return PgConnGracePeriod(grace)
 
 
@@ -14,8 +13,7 @@ def make(grace: float = 30.0) -> PgConnGracePeriod:
 # ---------------------------------------------------------------------------
 
 def test_negative_grace_period_falls_back_to_zero():
-    gp = PgConnGracePeriod(-5.0)
-    # after record_failure, process running, should act immediately
+    gp = PgConnGracePeriod(-5)
     with patch('src.pg_conn_grace_period.time') as t:
         t.time.return_value = 0.0
         gp.record_failure()
@@ -33,13 +31,12 @@ def test_reset_causes_should_act_true_without_failure():
 
 
 def test_reset_clears_failure_state():
-    gp = make(grace=60.0)
+    gp = make(grace=60)
     with patch('src.pg_conn_grace_period.time') as t:
         t.time.return_value = 0.0
         gp.record_failure()
         gp.reset()
         t.time.return_value = 1.0
-        # after reset, no failure context — acts normally
         assert gp.should_act(pg_running=True) is True
 
 
@@ -48,7 +45,7 @@ def test_reset_clears_failure_state():
 # ---------------------------------------------------------------------------
 
 def test_within_grace_pg_running_returns_false():
-    gp = make(grace=30.0)
+    gp = make(grace=30)
     with patch('src.pg_conn_grace_period.time') as t:
         t.time.return_value = 100.0
         gp.record_failure()
@@ -57,7 +54,7 @@ def test_within_grace_pg_running_returns_false():
 
 
 def test_within_grace_pg_not_running_returns_true():
-    gp = make(grace=30.0)
+    gp = make(grace=30)
     with patch('src.pg_conn_grace_period.time') as t:
         t.time.return_value = 100.0
         gp.record_failure()
@@ -70,7 +67,7 @@ def test_within_grace_pg_not_running_returns_true():
 # ---------------------------------------------------------------------------
 
 def test_expired_grace_pg_running_returns_true():
-    gp = make(grace=30.0)
+    gp = make(grace=30)
     with patch('src.pg_conn_grace_period.time') as t:
         t.time.return_value = 0.0
         gp.record_failure()
@@ -79,7 +76,7 @@ def test_expired_grace_pg_running_returns_true():
 
 
 def test_exactly_at_grace_boundary_returns_true():
-    gp = make(grace=30.0)
+    gp = make(grace=30)
     with patch('src.pg_conn_grace_period.time') as t:
         t.time.return_value = 0.0
         gp.record_failure()
@@ -92,7 +89,7 @@ def test_exactly_at_grace_boundary_returns_true():
 # ---------------------------------------------------------------------------
 
 def test_record_failure_keeps_first_timestamp():
-    gp = make(grace=30.0)
+    gp = make(grace=30)
     with patch('src.pg_conn_grace_period.time') as t:
         t.time.return_value = 0.0
         gp.record_failure()
@@ -107,7 +104,7 @@ def test_record_failure_keeps_first_timestamp():
 # ---------------------------------------------------------------------------
 
 def test_zero_grace_acts_immediately():
-    gp = make(grace=0.0)
+    gp = make(grace=0)
     with patch('src.pg_conn_grace_period.time') as t:
         t.time.return_value = 0.0
         gp.record_failure()
